requirements.txt
numpy
opencv-contrib-python
scipy
matplotlib
pillow
imutils
h5py
requests
progressbar2
scikit-learn
scikit-image
tensorflow[and-cuda]
# tensorflow-gpu
keras

